fix bug #472585: bad revision links in the directory view.
The problem was that I "optimized" some link generation. The mime type for
viewing was moved from within the code to the template. However, that mime
typed keyed download_url() to construct the url a bit differently. And that
alternate link generation was flat out busted. (note that setting
allow_markup to zero in the .conf file would have create the same broken
links)

This change just constructs the URL manually since the net effect of
download_url() is a very simple URL. This is much clearer than the use of
the function (or the prior use of download_link())