I've rewritten the handling of icons in 'standalone.py':  Currently only very few icons from the set of icons distributed with Apache are really used within the ViewCVS templates.  I've simply converted them into a tiny new module, which is imported and used by standalone.py.  Now standalone.py is really 'standalone' and can be run without Apache alltogether.  The former options '-i' and '--apache_root=' of standalone.py are now obsolete and I've removed them again.
The tool 'bin2inline_py.py' that I have used to convert the icon files
into a single sourcecode file is a cut-and-rape-oneway script from
something I've written a while ago to put icons into some other Tkinter
standalone program.  I case someone wants to add another icon, I've
placed this hacked version into the tools subdirectory and modified
'make-release', so that it will not be included into the final
ViewCVS 0.8 tarball.

For those who want to run standalone directly from the ViewCVS
source distribution, I've added a new option '-r' (or --repository=)
to standalone.py, so that it is now possible to patch the default
path '/home/cvsroot' without having to setup a Symlink in /home
pointing to their repository.  (Not everybody has the permissions
to do that on their server.)