Richard Offer posted Patch [ #498165 ] listen on non loopback interfaces: > The attached patch adds the ability for the standalone > server to listen on interfaces other than the loopback > (localhost). > >  usage: > >  ./standalone -h `hostname` > >  Will allow anyone not on the localmachine to still view > the cvs repo, without requiring a full-blown web server > be set up first. > > >  Patch is against release 0.9.1
I must admit, that I personally don't fully understand, why this
should be required.  But after adding some additional lines to the
gui() function and to the option initialization I applied this patch.
I tested it here with the name of my notebook and didn't see any
difference.  So at least it didn't hurt. ;-)