Finish issue #180 - Query results cache staleness under mod_python
Remove the code which attempts to cache database connection wrapper
objects.  That was useful long ago when the ViewVC code was sloppier,
but now the database handles get passed around in the call stack, so
there's a 1-to-1 mapping of ViewVC process to database connection.
The cache is, therefore, entirely unnecessary (and is causing
problems, in fact, for mod_python users).

* lib/cvsdb.py
  (gCheckinDatabase, gCheckinDatabaseReadOnly): Remove.
  (ConnectDatabase): Add 'readonly' optional parameter.  Don't try to
    play the connection caching game any more -- nobody cares.
  (ConnectDatabaseReadOnly): Reimplement as a wrapper around ConnectDatabase().