Get per-root overrides of authorizer stuffs working, and advertise this feature in viewvc.conf.
NOTE: Root listing views still currently suffer a bit here.  We can't
just overlay per-root options as we loop over roots because each root
will wind up with the union of the overrides for roots previously
handled in the loop.  We either need to find a way to clone the Config
object before overriding its values with per-root options (so the
original Config object is untouched) or find an elegant way to
*unoverlay* per-root options.  I suspect we'll have better luck with
the former approach.

* lib/viewvc.py
  (Request.run_viewvc): Overlap per-root options *before* setting up
    the authorizer for the root because, you know, the authorizer or
    its configuration might be what's being overridden.

* viewvc.conf.dist
  Note the existence of (and way to use) the per-root configuration overrides.