Allow root names to have a context
Presenting repositories as a flat list may e.g. lead to name clashes,
as was mentioned before (like in #53). This changeset allows root names
to contain slashes, with the purpose of presenting repositories in a
structured way. One way of using this, is simply by specifying a slash
in the 'svn_roots', 'cvs_roots' or 'renamed_roots' configuration. Another
way is with the extended 'root_parents' mechanism, where each root parent
can (optionally) be configured with a 'context' to use for the
parent's set of repositories.

This change is meant to be fully backwards compatible. In order to
prevent matches with valid (Windows) path names, the optional 'root
parents' context is specified with the '=' character. The 'root as url
component' mechanism has to look a bit harder for a match with a known
repository, but should still behave as before when no contexts are
used. Functionality has been tested with the standalone.py server, with
and without the 'root as url component' option and with and without the
'svnauthz' authorizer.