issue 126: Implement authorizer that uses Subversion own authz code
Implement a new ViewVC authorizer 'svn' which uses the Subversion
authorization API (via the Python bindings) to perform read access
checks.  Initially, this accepts configuration that is the same as
that of the existing 'svnauthz' authorizer, but has the potential to
support advanced syntaxes such as the glob support introduced in
Subversion 1.10.

Known shortcomings include the fact that the module grants root-level
access to everyone (while still denying reads of specific paths inside
a given repository, including the root directory), and that it is
unable to make a universal read access determination.

Patch by: Phil Richardson ppslim+viewvc{__AT__}gmail.com
          (Tweaked be me.)