Property names returned by repos.itemprops should be str.
Although Subversion claims all property name should be valid ascii subset of
XML 'name', there can be broken data not to obey this rule. So we should use
'surrogateescape' errors driver to decode them in vclib/, and we should
check if it is valid name viewvc.py to display them.