issue #214: Fix exception after requests on StandaloneServer
As the caller of viewvc.run_viewvc() does't expect that the server output
stream when the function returns.  However the text wrapper object created
in get_writeready_server_file() closed the stream when viewvc.run_viewvc()
returned, by the destructor of the wrapper object.  That's why
StandAloneServer reported ValueError Exception.  So we use custom wrapper
object not to close underlaid stream when the close() is called or the
object is removed.

* lib/viewvc.py
  (TextIOwrapper_noclose):
    New child class of io.TextIOWrapper. This never call close() on
    underlaid stream.
  (get_writeready_server_file):
    Use TextIOwrapper_noclose() instead of io.TextIOWrapper.