Fix issue #288: cmp() workaround cannot handle None input
Python2's `cmp()` function no longer exists in Python3, so when we
converted to Python3 we reproduced the seemingly relevant bits of
`cmp()` as a utility method.  Unfortunately, the naive suggestion from
Python's own upgrade docs:

    def cmp(a, b):
        return (a > b) - (a < b)

...is a bit *too* naive.  Python2's `cmp()` supports `None` input values as
well as comparisons between differing types of input values.  We don't
need the latter, but it seems in some cases we do need the former.  So
this change trades multiple replicas of the naive `cmp()` workaround for
a single shared slightly-less-naive implementation that handles `None`
inputs the way Python2's `cmp()` does.