Follow up to issue #288: Remove _logsort_date_cmp and _logsort_rev_cmp to reduce the use of cmp()
In bincvs.py and ccvs.py, we used _logsort_date_cmp() and
_logsort_rev_cmp() as cmp parameter for list.sort() method in Python 2,
for sorting CVS logs.  However it can be done by using key parameter for
list.sort(), with operator.attrgetter(), as "Sorting HOW TO"[1] document
shows.

So we use it and then remove _logsort_date_cmp and _logsort_rev_cmp
functions.

[1] https://docs.python.org/3/howto/sorting.html

* lib/vclib/ccvs/bincvs.py
  (_logsort_date_cmp, _logsort_rev_cmp): Removed.

* lib/vclib/ccvs/bincvs.py (BinCVSRepository.itemlog),
  lib/vclib/ccvs/ccvs.py (CCVSRepository.itemlog):
    Use operator.attrgetter() as key parameter for sorting.