Update very old comment, which has been out of context today.
The comment here was for the context that we got the file data here.
We get them before calling sort_file_data now, so the information
needs here is that there are such files that don't have revison info,
and it can be only in cvs repo. No functional change.

* lib/viewvc.py (sort_file_data)