Add new config variable 'options.repos_locale'
It determines the encoding of repository paths on the local system.
Also it can be appropriate for the default encoding of the file
contents in the repository.  However it can be the case that the
encoding of the repository path and the encoding of the most files'
content is different.  So we only change the default of config
variable 'options.default_encoding' from 'utf-8' to None, which
means using locale encoding.

* conf/viewvc.conf.dist
  (options.repos_locale):
    New.
  (options.default_encoding):
    Change default value and add comment for default value

* lib/config.py
  (Config.set_defaults):
    Set new default values for 'options.repos_locale' and
    'options.default_encoding'

* lib/common.py
  (locale_tuple_to_name, get_current_lc_ctype):
    New. Those are helper functions for get_repos_encodings().
  (get_repos_encodings):
    New.