issue #301: Fix content encoding handling on diff view and on patch.
In revision diff view, ViewVC should show the diff as UTF-8 encoded
Unicode HTML, by transcoding if the file contents is encoded in some
encoding other than UTF-8 or ASCII. On the other hand, with generate
patch mode, ViewVC should provide a raw diff without transcoding.

Those were not done because vclib._diff_fp(), which is used in
diff view without highlighting of intraline changes and generate
patch mode, always decoded the file content as UTF-8.

With this commit, we add a new argument to specify the encoding
to decode the diff stream, and make the code path to handle the
diff content without decoding/encoding, includes vclib._diff_fp().

To accomplish it, we also add a new argument "is_text" to rawdiff()
method in vclib.Repository base class and its discendants.

* lib/vclib/__init__.py
  (Repository.rawdiff):
    - Rename argument "type" to "diff_type", not to hide the "type"
      class.
    - Add new argument "is_text" to specify the stream returned
      by the method should be a decoded stream or raw bytes stream.
  (_diff_fp.__init__):
    Add new argument "encoding" which is used to decode (or not
    decode if it is None) the diff stream from diff utility.
  (_diff_fp.read, _diff_fp.readline, _check_process_errors):
    Fix to work with both of bytes and str stream.

* lib/vclib/ccvs/bincvs.py (BaseCVSRepository.rawdiff),
  lib/vclib/ccvs/ccvs.py (CCVSRepository.rawdiff),
  lib/vclib/svn/svn_ra.py (RemoteSubversionRepository.raw_diff),
  lib/vclib/svn/svn_repos.py (LocalSubversionRepository.rawdiff):
    Impliment the change of specification on base class.

* lib/viewvc.py
  (diff_parse_headers):
    Add new argument is_text to allow bytes stream for "fp".
  (view_patch):
    Handle the diff stream "fp" as bytes stream, without any
    decode/encode by Python.
  (DiffDescription._content_fp):
    Request diff stream as str explicitly.
  (DiffDescription._prop_fp):
    Request diff stream decoded with the repository's content encoding.