standalone.py: Fix htpasswd support
While processing "Authorization:" header, we decoded authorization
data encoded in BASE64 by using base64.b64decode(). This function
returns bytes both in Python 2 and in Python 3, and we treated it
as a str object. It was correct in Python 2, however it needs to
explicitly decode to str. This fixes it.

* bin/standalone.py (ViewVCHTTPRequestHandler.run_viewvc)