* lib/vclib/ccvs/rcsparse/test-data/*.out: Update test data for bytes stream.
When the rcsparse module was ported to Python3, it parsed RCS files
as text stream,  and then for parsing RCS files for binary content,
it fixed to be parsing them as bytes stream.
(eef060eafad369296f15453ed3d754b5b892966d)
This commit is its follow up, updating expected output of the tests.