issue 385: Drop dependency on the cgi module
As has been done many times in the past, implement a (scope-reduced)
compatibility shim for cgi.parse() in order to avoid introducing
another third party dependency.  Should we need a fuller CGI interface
in the future, we can revisit this decision, but for now let's favor
simplicity for ViewVC admins.